requirements.txt
langchain
streamlit
pandas
python-dotenv
langchain-experimental
tabulate
openai
langchain_openai